requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
bcrypt==4.0.1
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3